Former-commit-id: e4af7e2ef3af5a32f2cc9e5a5c905d5d4c2a74c9

+++ cave/com.raytheon.uf.viz.damagepath/src/com/raytheon/uf/viz/damagepath/ExportDamagePathAction.java
@@ -21,12 +21,14 @@ package com.raytheon.uf.viz.damagepath;
 
 import java.io.FileOutputStream;
 
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.FileDialog;
 import org.eclipse.swt.widgets.Shell;
+import org.geotools.data.simple.SimpleFeatureCollection;
 
-import com.raytheon.uf.common.json.geo.GeoJsonUtil;
-import com.raytheon.uf.common.json.geo.GeoJsonUtilSimpleImpl;
+import com.raytheon.uf.common.json.geo.IGeoJsonService;
+import com.raytheon.uf.common.json.geo.SimpleGeoJsonService;
 import com.raytheon.uf.common.status.IUFStatusHandler;
 import com.raytheon.uf.common.status.UFStatus;
 import com.raytheon.uf.viz.core.VizApp;
@@ -42,7 +44,13 @@ import com.raytheon.viz.ui.cmenu.AbstractRightClickAction;
  * 
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
- * Feb 9, 2015  3975       njensen     Initial creation
+ * Feb 09, 2015  3975      njensen     Initial creation
+ * Apr 23, 2015  4354      dgilling    Export as GeoJSON Feature object.
+ * Jun 05, 2015  4375      dgilling    Prompt user before exporting feature 
+ *                                     with no polygons.
+ * Jun 09, 2015  4355      dgilling    Rename action for UI.
+ * Jun 18, 2015  #4354     dgilling    Support FeatureCollections so each 
+ *                                     polygon can have its own properties.
  * 
  * </pre>
  * 
@@ -55,10 +63,14 @@ public class ExportDamagePathAction extends AbstractRightClickAction {
     protected static final transient IUFStatusHandler statusHandler = UFStatus
             .getHandler(ExportDamagePathAction.class);
 
+    private static final String CONFIRM_DLG_MSG = "Damage Path has no polygons. Continue with exporting file?";
+
+    private static final String CONFIRM_DLG_TITLE = "Damage Path has no polygons!";
+
     protected static final String[] EXTENSIONS = new String[] { "*.json" };
 
     public ExportDamagePathAction() {
-        super("Export GeoJSON");
+        super("Export to File");
     }
 
     @Override
@@ -74,9 +86,22 @@ public class ExportDamagePathAction extends AbstractRightClickAction {
 
                 if (filename != null) {
                     DamagePathLayer<?> layer = (DamagePathLayer<?>) getSelectedRsc();
+                    SimpleFeatureCollection featureCollection = layer
+                            .buildFeatureCollection();
+
+                    if (featureCollection.size() < 1) {
+                        boolean export = MessageDialog.openConfirm(shell,
+                                CONFIRM_DLG_TITLE, CONFIRM_DLG_MSG);
+                        if (!export) {
+                            statusHandler
+                                    .info("User chose to cancel exporting damage path file because it had no polygons.");
+                            return;
+                        }
+                    }
+
                     try (FileOutputStream fos = new FileOutputStream(filename)) {
-                        GeoJsonUtil json = new GeoJsonUtilSimpleImpl();
-                        json.serialize(layer.getPolygon(), fos);
+                        IGeoJsonService json = new SimpleGeoJsonService();
+                        json.serialize(featureCollection, fos);
                     } catch (Exception e) {
                         statusHandler.error(
                                 "Error exporting damage path file to "
